requirements.txt
Flask==2.1.2
Flask-MySQLdb==1.0.1
mysql-connector-python
pandas
scikit-learn
